# update setup_gcsfuse for better perf